requirements/dev.txt
python-dotenv==1.0.1 
aiohttp==3.9.3
aiokafka==0.10.0
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
confluent-kafka==2.3.0
discord.py==2.3.2
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
kafka==1.3.5
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sympy==1.12
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.37.2
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
torch==2.2
